requirements.txt
webex_bot==0.3.4
coloredlogs==15.0.1
openai
pyyaml
